# backend/app/routers/invites.py
import secrets

from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, status
from sqlmodel import select

from app.database.database import SessionDep
from app.database.models.group import Group
from app.database.models.invite import (
    GroupInvite,
    GroupInviteCreate,
    GroupInviteResponse,
    InvitationTokenResponse,
)
from app.database.models.user import User
from app.database.models.users_groups import UsersGroups
from app.dependencies.auth import CurrentUser
from app.middleware.is_user_group import GroupMembership
from app.services.auth import oauth2_scheme
from app.services.email import EmailService

router = APIRouter(
    prefix="/invites",
    tags=["invites"],
    dependencies=[Depends(oauth2_scheme)],
)


@router.get(
    "/my-invites",
    response_model=list[GroupInviteResponse],
    status_code=status.HTTP_200_OK,
)
async def get_my_invites(session: SessionDep, current_user: CurrentUser) -> list[GroupInviteResponse]:
    invites = session.exec(
        select(GroupInvite).where(GroupInvite.email == current_user.email)
    ).all()

    return [
        GroupInviteResponse(
            id=invite.id,
            email=invite.email,
            group_id=invite.group_id,
            token=invite.token,
            created_at=invite.created_at,
        )
        for invite in invites
    ]


@router.post(
    "/{group_id}/generate",
    response_model=InvitationTokenResponse,
    status_code=status.HTTP_201_CREATED,
)
async def generate_invite_link(
    session: SessionDep,
    db_group: GroupMembership,
) -> InvitationTokenResponse:
    token = secrets.token_urlsafe(16)

    invite = GroupInvite(group_id=db_group.id, token=token)

    session.add(invite)
    session.commit()
    session.refresh(invite)

    return InvitationTokenResponse(token=token)


@router.post(
    "/{group_id}/email",
    response_model=GroupInviteResponse,
    status_code=status.HTTP_201_CREATED,
)
async def invite_by_email(
    invite_data: GroupInviteCreate,
    session: SessionDep,
    db_group: GroupMembership,
    background_tasks: BackgroundTasks,
) -> GroupInviteResponse:
    if not invite_data.email:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Email is required for email invitations",
        )

    user_exists = session.exec(select(User).where(User.email == invite_data.email)).first()

    if not user_exists:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")

    is_member = session.exec(
        select(UsersGroups).where(
            UsersGroups.user_id == user_exists.id,
            UsersGroups.group_id == db_group.id,
        )
    ).first()

    if is_member:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="User is already a member of this group",
        )

    existing_invite = session.exec(
        select(GroupInvite).where(
            GroupInvite.email == invite_data.email,
            GroupInvite.group_id == db_group.id,
        )
    ).first()

    if existing_invite:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="An invitation has already been sent to this email",
        )

    token = secrets.token_urlsafe(16)

    invite = GroupInvite(email=invite_data.email, group_id=db_group.id, token=token)

    session.add(invite)
    session.commit()
    session.refresh(invite)

    background_tasks.add_task(
        EmailService.send_group_invite, invite_data.email, db_group.name, token
    )

    return GroupInviteResponse(
        id=invite.id,
        email=invite.email,
        group_id=invite.group_id,
        token=invite.token,
        created_at=invite.created_at,
    )


@router.put("/accept/{token}", response_model=dict, status_code=status.HTTP_200_OK)
async def accept_invite(
    token: str,
    session: SessionDep,
    current_user: CurrentUser,
) -> dict:
    invite = session.exec(select(GroupInvite).where(GroupInvite.token == token)).first()

    if not invite:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Invalid invitation")

    if invite.email and invite.email != current_user.email:
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="This invitation is for a different user",
        )

    is_member = session.exec(
        select(UsersGroups).where(
            UsersGroups.user_id == current_user.id,
            UsersGroups.group_id == invite.group_id,
        )
    ).first()

    if is_member:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="You are already a member of this group",
        )

    group = session.exec(select(Group).where(Group.id == invite.group_id)).first()

    if not group:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Group not found")

    users_groups = UsersGroups(user_id=current_user.id, group_id=group.id)

    if invite.email:
        session.delete(invite)

    session.add(users_groups)
    session.commit()

    return {"message": "Invitation accepted successfully"}


@router.delete("/reject/{token}", status_code=status.HTTP_204_NO_CONTENT)
async def reject_invite(token: str, session: SessionDep, current_user: CurrentUser):
    invite = session.exec(select(GroupInvite).where(GroupInvite.token == token)).first()

    if not invite:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Invalid invitation")

    email_matches_user = invite.email and invite.email == current_user.email
    invite_belongs_to_user_in_group = invite.group_id in [group.id for group in current_user.groups]

    if not email_matches_user and not invite_belongs_to_user_in_group:
        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Invalid invitation")

    session.delete(invite)
    session.commit()